+++ src/main/java/blackjack/Blackjack.java
@@ -7,6 +7,10 @@ import blackjack.domain.common.Names;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.GamePlayer;
 import blackjack.domain.player.Players;
+import blackjack.domain.result.DealerResult;
+import blackjack.domain.result.GamePlayerResult;
+import blackjack.domain.result.Result;
+import java.util.ArrayList;
 import java.util.List;
 
 public class Blackjack {
@@ -26,6 +30,17 @@ public class Blackjack {
                                             .map(name -> new GamePlayer(name, drawTwo()))
                                             .toList();
         return new Players(dealer, gamePlayers);
+    }
+
+    public Result checkResult(Dealer dealer, List<GamePlayer> gamePlayers) {
+        List<GamePlayerResult> gamePlayerResults = new ArrayList<>();
+
+        for (GamePlayer gamePlayer : gamePlayers) {
+            gamePlayerResults.add(
+                    new GamePlayerResult(gamePlayer.getName(), dealer.checkPlayer(gamePlayer)));
+        }
+
+        return new Result(gamePlayerResults, DealerResult.of(dealer.getName(), gamePlayerResults));
     }
 
     private Cards drawTwo() {
