+++ src/test/java/blackjack/BlackjackTest.java
@@ -2,9 +2,16 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.CardValue;
+import blackjack.domain.card.Cards;
+import blackjack.domain.common.Name;
 import blackjack.domain.common.Names;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.GamePlayer;
 import blackjack.domain.player.Player;
+import blackjack.domain.result.ResultStatus;
 import blackjack.fixture.CardFixture;
+import blackjack.fixture.PlayerFixture;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -25,6 +32,24 @@ class BlackjackTest {
                            .get(0), 13);
         assertPlayer(result.getGamePlayers()
                            .get(1), 15);
+    }
+
+    @Test
+    @DisplayName("블랙잭은 게임 결과를 종합한다.")
+    public void Dealer_Count_result() {
+        GamePlayer gamePlayer = PlayerFixture.게임_플레이어_생성(List.of(CardValue.EIGHT, CardValue.THREE));
+        Name name = new Name("딜러");
+        Cards cards = CardFixture.카드_목록_생성(List.of(CardValue.EIGHT, CardValue.FOUR));
+        Dealer dealer = new Dealer(name, cards);
+        var sut = new Blackjack();
+
+        var result = sut.checkResult(dealer, List.of(gamePlayer));
+
+        assertThat(result.getGamePlayerResults()
+                         .get(0)
+                         .getResultStatus()).isEqualTo(ResultStatus.LOSE);
+        assertThat(result.getDealerResult()
+                         .getResultWithResultStatus(ResultStatus.WIN)).isEqualTo(1);
     }
 
     private void assertPlayer(Player player, int value) {
